feat: Add support for Azure OpenAI embeddings and model deployment
This commit adds support for specifying the deployment name for Azure OpenAI embeddings and model in the configuration file. The user is prompted to enter the deployment name for both embeddings and model if the corresponding host is set to Azure-OpenAI. The entered deployment names are then stored in the config dictionary as "embeddings-deployment" and "model-deployment" respectively.