Refactor code to include caching and syncing functionality
This commit adds caching and syncing functionality to the code. It introduces a new `cache.py` module that includes functions for loading and saving vector caches, as well as getting the cache path and creating the cache directory. The `VectorCache` class is also added to represent the vector cache data structure.

In the `app.py` module, the code is modified to use the new cache functions. The `save_vector_cache` function is called to save the vector cache after parsing the codebase and syncing the documents. The `load_vector_cache` function is used to load the vector cache if it exists.

In the `codeparser.py` module, the `commit_hash` is extracted using the `repo.get_commit_hash` function and added to the metadata of each document.

In the `repo.py` module, the `get_commit_hash` function is added to retrieve the commit hash of a file using the `git log` command.

In the `vector_store.py` module, the `VectorStore` class is modified to handle caching and syncing. The `vector_cache` attribute is added to store the vector cache. The `sync_documents` method is implemented to sync the vector store with the latest documents, updating the cache